requirements.txt
numpy
pandas
transformers
torch
biopython
accelerate
einops
pyyaml
tqdm
